refactor: broke out deduplicateConcert function to its on file
