+++ components/lists/ConcertGrid.tsx
@@ -11,6 +11,7 @@ import { ITicketmasterSearchResponse } from "../../types/ITicketmasterEvent";
 import { IConcertCard } from "../../types/IConcertCard";
 import { ActivityIndicator } from "react-native-paper";
 import { IGenreName } from "../../types/IGenreName";
+import { deduplicateConcerts } from "../../utils/deduplicateConcerts";
 
 interface ConcertGridProps {
   selectedFilters: string[];
@@ -68,22 +69,11 @@ export default function ConcertGrid({ selectedFilters }: ConcertGridProps) {
     }
   });
 
-  const allConcerts: IConcertCard[] =
-    concertList?.pages
-      .flatMap((page) => (page._embedded?.events ?? []).map(mapToConcertCard))
-      .reduce<IConcertCard[]>((acc, current) => {
-        const exists = acc.some(
-          (c) =>
-            c.artist === current.artist &&
-            new Date(c.date).toDateString() ===
-              new Date(current.date).toDateString()
-        );
-        if (!exists) acc.push(current);
-        return acc;
-      }, [])
-      .sort(
-        (a, b) => new Date(a.date).getTime() - new Date(b.date).getTime()
-      ) ?? [];
+  const allConcerts: IConcertCard[] = deduplicateConcerts(
+    concertList?.pages.flatMap((page) =>
+      (page._embedded?.events ?? []).map(mapToConcertCard)
+    ) ?? []
+  );
 
   const handleEndReached = () => {
     if (hasNextPage) {
