+++ utils/deduplicateConcerts.ts
@@ -1,0 +1,16 @@
+import { IConcertCard } from "../types/IConcertCard";
+
+export function deduplicateConcerts(concerts: IConcertCard[]): IConcertCard[] {
+  return concerts
+    .reduce<IConcertCard[]>((acc, current) => {
+      const exists = acc.some(
+        (c) =>
+          c.artist === current.artist &&
+          new Date(c.date).toDateString() ===
+            new Date(current.date).toDateString()
+      );
+      if (!exists) acc.push(current);
+      return acc;
+    }, [])
+    .sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime());
+}
